feat: decouple grid2d-specific imports for multi-backend support (ROADMAP #9)
Move shared map types (MapInfo, load_map_info, OBSTACLE_SPACE, FREE_SPACE)
from navirl.backends.grid2d to navirl.core.maps so that metrics, validators,
visualization, and layout modules can work with any backend. Update scenario
validation to accept any registered backend name instead of hardcoding
"grid2d". Update base_env.py to use the plugin registry instead of directly
importing Grid2DBackend. Add dt property to SceneBackend ABC. Fix lint issues
(import sorting, unused variable).
