+++ navirl/viz/render.py
@@ -8,7 +8,7 @@ import cv2
 import numpy as np
 import yaml
 
-from navirl.backends.grid2d.maps import load_map_info
+from navirl.core.maps import load_map_info
 
 
 def _world_to_px(
