+++ tests/test_judge.py
@@ -19,7 +19,6 @@ from navirl.verify.judge import (
     write_judge_output,
 )
 
-
 # ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------
