+++ tests/test_validators.py
@@ -35,7 +35,6 @@ from navirl.verify.validators import (
     validate_units_metadata,
 )
 
-
 # ---------------------------------------------------------------------------
 # Fixtures
 # ---------------------------------------------------------------------------
